third_party/WebKit/LayoutTests/fast/events/platform-wheelevent-paging-y-in-scrolling-div-expected.txt

PASS event.wheelDeltaY is window.givenScrollTop * -3
PASS event.wheelDeltaX is 0
PASS event.wheelDelta is window.givenScrollTop * -3
PASS div.scrollTop == window.expectedScrollTop && div.scrollLeft == window.expectedScrollLeft became true
PASS successfullyParsed is true

TEST COMPLETE
